+++ Tests/VCoreTests/Tests/Extensions/Foundation/ArrayFirstAndLastInstanceTests.swift
@@ -1,5 +1,5 @@
 //
-//  ArrayFirstAndLastInstancesTests.swift
+//  ArrayFirstAndLastInstanceTests.swift
 //  VCoreTests
 //
 //  Created by Vakhtang Kontridze on 29.11.22.
@@ -9,7 +9,7 @@ import XCTest
 @testable import VCore
 
 // MARK: - Tests
-final class ArrayFirstAndLastInstancesTests: XCTestCase {
+final class ArrayFirstAndLastInstanceTests: XCTestCase {
     // MARK: Test Data
     private let array: [any P] = [
         S1(value: 1),